ranks_test_task/requirements.txt
asgiref==3.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Django==4.0.4
django-rest-framework==0.1.0
djangorestframework==3.13.1
Flask==2.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
stripe==3.0.0
urllib3==1.26.9
Werkzeug==2.1.2
